Add ng2-chart.js an embedded video
Ng2-chart.js was added in order to test if it allows multiple
different-element-size lines in one chart. The resulting chart
does not go accordingly to the User_Story_02(homogenization of
lines chats). On the other hand, It is still in research stage
how to embedded private YOUTUBE videos inside the web.