Creatd  PUT method for kinematic API
The kinematic controller did not hav thee put method, so It  was no
possible to add or update new data to an specific kinematic analysis
after being created. This implemented API allows the input data to
be cooked in order to feed the Highchart object which is  embeded
in the controller to later hit the database. The data which is send
to  database belong to the left gait analysis only. The impact is
that highcharts graphics are feeded with real data that belong to the
patient; visually speaking , graphic results comes to live in the
user interface.