Hot-Fix Hidded Kinematic Analysis preview when teraphist or patient changed
When a certain kinematic analysis is selected from list after selec-
ting a thearpist and a patient, the kinematic result preview UI  did
not changed or disapeared if the current  therapist or patient changed
which caused confusions the final user. This was fixed by hidding
whatever kinematic analysis is selected after changing the current
therapist or patient seleced in the parent component. The impact of
is to make a coherent kinematic analysis preview view  belong to the
current therapist and pattient selected.