Added status logic of kKinematicAnalysis
The status of the kinematic analysis was not stablished since its
creation to its completeness. The solution implemented was to add
an array of states  everytime the kinematic analysis suffered a
change, specifying what the new status was based on the bussines
logi. The impact of this the kinematic analysis entity works in
two ways: as an Appointment and as a kinematic analysis.