Added Homepage Contact_us functionality
Contact_us section in homepage had UI only, no front-end and backend
implemented. This did not allow customers to leave their contact info
in our servers. From front-end side, bindings were implemented accor-
dig business logic. From backend side, It was implemented the model,
the controller, and the api in regards to contact functionality.
Api integration was done as well between the front-end and backend.
The impact of this implementation is that customers that are inter-
terested in gaitcome --software/hardware as a service-- can store
their contact info is gaitcome servers.