Added Google Analitycs
Google analytics was integrated plus minor changes for demo access.
This will allow the application to be tracked by google analytics
to have a better understanding of customers behavior. The impact
of this implementation is that gaitcome will be able to know where
and what functionalities to improve in shor time.