Added Angular_Yarn_Sass_Upgrade
After adding a SSL certificate, gaitcome time loading pages are
extremtely slow. Possible reason is the dist folder which contians
a malfunctioning bundle generated by angular client version 4.

Due to new freatures was realased in angular 9 related to the
current issue (time loading of the web page), It was decided
to migrate to this version.

Additionally, npm pakage manageris being replaced by Yarn package
manager because of its faster way to manage install and maintain
packages.

The last but not the least, Sass is now used in the proyect for it
is easy to use nest style elements to be rendered in html.